% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcutils.R
\name{getdims}
\alias{getdims}
\title{Get dimensions of counts, data and scale.data slots in all the assays in a seurat object}
\usage{
getdims(obj)
}
\arguments{
\item{obj}{A Seurat Object}
}
\value{
A dataframe printed on the console
}
\description{
This function returns a dataframe listing dimensions for counts, data and scale.data slot
in RNA and other present assays
}
