% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcutils.R
\name{getplots}
\alias{getplots}
\title{Get a list of dimension-plots}
\usage{
getplots(sro, sr, dd, plt.title)
}
\arguments{
\item{sro}{A Seurat Object}

\item{sr}{A vector of cell barcodes to highlight. These are the barcodes predicted to be doublets by Scrublet}

\item{dd}{A vector of cell barcodes to highlight. These are the barcodes predicted to be doublets by DoubletDecon}

\item{plt.title}{A title for the plots}
}
\value{
A list of three plots
}
\description{
This function returns a list of 3 dimension plots for your single cell data.
The first plot is a basic DimPlot output
The second plot is a DimPlot with cells.highlighted with the cells provided
The third plot is a DimPlot with cells.highlighted with the cells provided
}
