% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcutils.R
\name{seurat2cloupe}
\alias{seurat2cloupe}
\title{Extract data to import to Loupe Cell Browser}
\usage{
seurat2cloupe(object, reduction, dims, metadata, keyword, opdir)
}
\arguments{
\item{object}{A Seurat Object}

\item{reduction}{The reduction to extract the coordinates for. 'tsne' or 'umap'}

\item{dims}{The number of dimensions to extract for a give reduction}

\item{metadata}{The column name in the seurat object metadata table to extract}

\item{keyword}{A keyword for the filename}

\item{opdir}{The output directory in which the files will be written}
}
\value{
Two .csv files written to the specified output directory - data4cloupe_keyword.csv contains the reduction coordinates and cluster4cloupe_keyword.csv contains the metadata information.
}
\description{
This function extracts data from a given seurat object for import into Loupe Cell Browser. 
It generates two csv files - one containing the coordinates for all the cells for the reduction of interest 
and one containing meta-data information such as clusters, categories, etc.
}
